requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
langchain
langchain-community
langchain-google-genai
faiss-cpu
google-generativeai
apscheduler
